Added UUID system to client, order placing/cancelling
NOTE: when cancelling orders partially (i.e, only a set quantity being removed), the parameter "volume" is what the order quantity is SET to, not decremented (for some reason). This is the convention we will have to follow. Also when volume=0, it is a full cancel (the order is removed).