Restructured logging to general BookLogger and TransactionLogger objects
The motivation for this is so that in addition to the exchange server having logger objects, the clients can too. This will come in handy when the clients manage their own copies of the market history.

The main difference from the previous system is that instead of calling logging.info() on the logger objects, now an update_log() is called on the Book/TransactionLogger objects