Implemented Client-side book & transaction logging
Current system will create two txt files per client in the market_client/market_logs directory, with a name corresponding to the client id. This setup will probably need to be changed later (maybe a directory for each client's set of logs?).

TODO: Currently the client will only have a history of the market as far back as the moment the joined the exchange server. The exchange server should share with each new client its current book&transaction logs when they join, so that they can copy the data to their end.